Benchmark all backends on WNAT_Hagen: Python (5.63s), Rust (0.334s), C++ (0.091s)
- Fixed Rust pybind11 binding issue via clean maturin rebuild
- Fixed Python signed_area elem_ids dtype cast (np.intp)
- Measured full init on WNAT_Hagen (52,774 verts, 98,365 elems)
- Updated README table: v0.2.0 MATLAB (7.7s) → v0.4.1 Python (5.63s) → v0.5.0 Rust (0.334s) → v0.6.0 C++ (0.091s)
- 84.6× speedup from MATLAB to C++ backend

+++ src/chilmesh/CHILmesh.py
@@ -397,7 +397,7 @@ class CHILmesh(CHILmeshPlotMixin):
             elem_ids = np.arange( self.n_elems )
         if np.isscalar( elem_ids ):
             elem_ids = np.array( [elem_ids] )
-        elem_ids = np.asarray(elem_ids)
+        elem_ids = np.asarray(elem_ids, dtype=np.intp)
 
         if self.connectivity_list.shape[1] == 3:
             # Pure triangular mesh — single vectorised path
